% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcDistance.R
\name{abcDistance}
\alias{abcDistance}
\title{Calculate ABC Distances}
\usage{
abcDistance(summaryValuesMatrix, originalSummaryValues, pls.model.list)
}
\arguments{
\item{summaryValuesMatrix}{Matrix of summary statistics from simulations}

\item{originalSummaryValues}{Summary statistics for original data.}

\item{pls.model.list}{indexing of a list of pls models for different params}
}
\value{
Returns euclidean distance of each simulation's summary values to
the original summary stats.
}
\description{
This function uses results from a PLS model to calculate the distance from
each sim to the original data.
}
\details{
This function runs a PLS regression on each free parameter in the model, unlike
a true multivariate PLS regression.  For ABC, this seems to result in much
better results, without one parameter dominating the combined variance.
}
\examples{

data(simRun)

abcDistance(summaryValuesMatrix, originalSummaryValues, pls.model.list)

}
\author{
Brian O'Meara and Barb Banbury
}
