% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcparticle.R
\name{abcparticle}
\alias{abcparticle}
\title{abcparticle}
\usage{
abcparticle(id = NA, generation = NA, weight = NA, distance = NA,
  startingValues = NA, intrinsicValues = NA, extrinsicValues = NA)
}
\arguments{
\item{id}{Simulation ID}

\item{generation}{Simulation generation}

\item{weight}{Simulation weight}

\item{distance}{Simulation distance}

\item{startingValues}{Vector of parameter estimates for startingStates}

\item{intrinsicValues}{Vector of parameter estimates for intrinsicValues}

\item{extrinsicValues}{Vector of parameter estimates for extrinsicValues}
}
\description{
An internal TreEvo function that creates a list of objects in class
abcparticle
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}

