\name{abctaxon}
\alias{abctaxon}

\title{abctaxon}

\description{Internal TreEvo function to create a list of objects}

\usage{
abctaxon(id=0, name="taxon0", timeSinceSpeciation=0, states=NA, nextstates=NA)
}

\arguments{
  \item{id}{Simulation ID}
  \item{name}{Taxon of interest}
  \item{timeSinceSpeciation}{Time since speciation}
  \item{states}{States of taxon}
  \item{nextstates}{States of taxon after mutation}
}


\references{O'Meara and Banbury, unpublished}
\author{Brian O'Meara and Barb Banbury}

