% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcmodels.intrinsic.R
\name{autoregressiveIntrinsic}
\alias{autoregressiveIntrinsic}
\title{Intrinsic Character Evolution Models}
\usage{
autoregressiveIntrinsic(params, states, timefrompresent)
}
\arguments{
\item{params}{describes input paramaters for the model.
\code{autoregressiveIntrinsic} params = sd (sigma), attractor (character
mean), attraction (alpha)}

\item{states}{vector of states for each taxon}

\item{timefrompresent}{which time slice in the tree}
}
\value{
A matrix of values representing character displacement from a single
time step in the tree.
}
\description{
This function describes a model of intrinsic character evolution. New
character values are generated after one time step via a discrete-time OU
process.
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}
\keyword{autoregressiveIntrinsic}
\keyword{intrinsic}

