% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesCoverageProb.R
\name{bayesCoverageProb}
\alias{bayesCoverageProb}
\title{Bayesian Coverage Probability}
\usage{
bayesCoverageProb(RealParam, HPD, verbose = F)
}
\arguments{
\item{RealParam}{Real parameter values}

\item{HPD}{List of HPD from doRun_rej or doRun_prc results}

\item{verbose}{Commented screen output}
}
\value{
Returns a value for each free parameter that describes the
percentage the real value falls within the HPD
}
\description{
This function calculates coverage probability for a list of HPDs
}
\details{
Only for use with simulated data to test models.
}
\author{
Brian O'Meara and Barb Banbury
}
