% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareListIPD.R
\name{compareListIPD}
\alias{compareListIPD}
\title{Plot Interparticle Distance between Runs}
\usage{
compareListIPD(particleDataFrame, verbose = F)
}
\arguments{
\item{particleDataFrame}{A \code{particleDataFrame} object output from \code{doRun}}

\item{verbose}{Commented screen output.}
}
\value{
Produces a plot with IPD between runs per generation.
}
\description{
This function plots interparticle distance (IPD) between runs for each free
parameter.
}
\author{
Brian O'Meara and Barb Banbury
}
