% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTaxonFrameToGeigerData.R
\name{convertTaxonFrameToGeigerData}
\alias{convertTaxonFrameToGeigerData}
\title{Converting Simulated Data to Geiger Formatted Data}
\usage{
convertTaxonFrameToGeigerData(taxonframe, phy)
}
\arguments{
\item{taxonframe}{data frame with species values}

\item{phy}{Tree (ape phylo format)}
}
\value{
Returns a data frame in geiger format
}
\description{
This function takes simulated data and converts it to geigers formatting.
}
\author{
Brian O'Meara and Barb Banbury
}
