% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonDFWithPossibleExtinction.R
\name{createAbcTaxonFromHeightsRow}
\alias{createAbcTaxonFromHeightsRow}
\title{Create abctaxon object}
\usage{
createAbcTaxonFromHeightsRow(x)
}
\arguments{
\item{x}{A row of the heights df}
}
\value{
An abctaxon object
}
\description{
This creates an abctaxon object from a row in the heights df
created within getAbcTaxonListWithPossibleExtinction
}
\author{
Brian O'Meara
}
