% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credibleInt.R
\name{credibleInt}
\alias{credibleInt}
\title{Bayesian Credible Interval}
\usage{
credibleInt(particleDataFrame, percent = 0.95)
}
\arguments{
\item{particleDataFrame}{A \code{particleDataFrame} object, as found among the output from \code{\link{doRun}} functions.}

\item{percent}{Probability content of the highest probability density.}
}
\value{
Returns a matrix with weighted mean, standard deviation, upper and lower credible
intervals for each free parameter.
}
\description{
This function calculates credible interval for each free parameter.
}
\examples{

data(simRunExample)
credibleInt(results[[1]]$particleDataFrame)

}
\author{
Brian O'Meara and Barb Banbury
}
