% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSimulationForPlotting.R
\name{doSimulationForPlotting}
\alias{doSimulationForPlotting}
\title{Discrete Time Character Simulation}
\usage{
doSimulationForPlotting(splits, intrinsicFn, extrinsicFn, startingValues,
  intrinsicValues, extrinsicValues, timeStep, plot = FALSE,
  savePlot = FALSE, saveHistory = FALSE, saveRealParams = FALSE,
  jobName = "")
}
\arguments{
\item{intrinsicFn}{Name of intrinsic function characters should be simulated
under}

\item{extrinsicFn}{Name of extrinsic function characters should be simulated
under}

\item{startingValues}{State at the root}

\item{intrinsicValues}{Vector of values corresponding to the params of the
intrinsic model}

\item{extrinsicValues}{Vector of values corresponding to the params of the
extrinsic model}

\item{timeStep}{This value corresponds to the number of discrete time steps
on the shortest branch}

\item{plot}{Will create a new interactive window that plots character values
throughout the history of the tree}

\item{savePlot}{Saves the character tree using jobName}

\item{saveHistory}{Saves the character history throughout the simulation}

\item{saveRealParams}{Saves intrinsicValues and extrinsicValues as both a
matrix and a vector}

\item{jobName}{Optional name for the job}

\item{phy}{A phylogenetic tree, in package \code{ape}'s \code{phylo} format.}
}
\value{
A data frame of species character (tip) values in the tree.
}
\description{
This function evolves continuous characters in a discrete time process
}
\details{
When saveHistory is TRUE, processor time will increase quite a bit.
SaveRealParams is useful for tracking the "real" true values if simulating
data for abc runs.  It is not useful for empirical abc runs.
}
\examples{


tree<-rcoal(30)

#Simple Brownian motion
char<-doSimulationForPlotting(
	phy=tree,
	intrinsicFn=brownianIntrinsic,
	extrinsicFn=nullExtrinsic,
	startingValues=c(10), #root state
	intrinsicValues=c(0.01),
	extrinsicValues=c(0),
	timeStep=0.0001,
	plot=FALSE,
	saveHistory=FALSE)


#Character displacement model with minimum bound
char<-doSimulationForPlotting(
	phy=tree,
	intrinsicFn=boundaryMinIntrinsic,
	extrinsicFn=ExponentiallyDecayingPushExtrinsic,
	startingValues=c(10), #root state
	intrinsicValues=c(0.05, 10, 0.01),
	extrinsicValues=c(0, .1, .25),
	timeStep=0.001,
	plot=TRUE,
	saveHistory=FALSE)

}
\author{
Brian O'Meara and Barb Banbury
}
