% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSimulationWithPossibleExtinction.R
\name{doSimulationWithPossibleExtinction}
\alias{doSimulationWithPossibleExtinction}
\title{Discrete Time Character Simulation}
\usage{
doSimulationWithPossibleExtinction(phy, intrinsicFn, extrinsicFn,
  startingValues, intrinsicValues, extrinsicValues, timeStep,
  saveHistory = FALSE, saveRealParams = FALSE, jobName = "",
  returnAll = FALSE, verbose = FALSE, reject.NaN = TRUE)
}
\arguments{
\item{phy}{A phylogenetic tree, in package \code{ape}'s \code{phylo} format.}

\item{intrinsicFn}{Name of intrinsic function characters should be simulated
under}

\item{extrinsicFn}{Name of extrinsic function characters should be simulated
under}

\item{startingValues}{State at the root}

\item{intrinsicValues}{Vector of values corresponding to the params of the
intrinsic model}

\item{extrinsicValues}{Vector of values corresponding to the params of the
extrinsic model}

\item{timeStep}{This value corresponds to the number of discrete time steps
on the shortest branch}

\item{saveHistory}{Saves the character history throughout the simulation}

\item{saveRealParams}{Saves intrinsicValues and extrinsicValues as both a
matrix and a vector}

\item{jobName}{Optional name for the job}

\item{returnAll}{If true, returns the values at each node}

\item{verbose}{If TRUE, chat about how the sim is going}

\item{reject.NaN}{Stop run if any simulated value is NaN}
}
\value{
A data frame of species character (tip) values in the tree (unless returnAll==TRUE, in which case it returns the raw df from the sim).
}
\description{
This simulation function evolves continuous characters in a discrete time process.
}
\details{
When \code{saveHistory} is \code{TRUE}, processor time will increase quite a bit.
\code{SaveRealParams} is useful for tracking the "real" true values if simulating
data for abc runs.  It is not useful for empirical abc runs.
}
\examples{

tree<-rcoal(30)

#Simple Brownian motion
char<-doSimulationWithPossibleExtinction(
	phy=tree,
	intrinsicFn=brownianIntrinsic,
	extrinsicFn=nullExtrinsic,
	startingValues=c(10), #root state
	intrinsicValues=c(0.01),
	extrinsicValues=c(0),
	timeStep=0.0001,
	saveHistory=FALSE)

}
\author{
Brian O'Meara and Barb Banbury
}
