% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitContinuous.hacked.R
\name{fitContinuous.hacked}
\alias{fitContinuous.hacked}
\title{Hacked version of Geiger's fitContinuous}
\usage{
fitContinuous.hacked(phy, data, data.names = NULL, model = c("BM", "OU",
  "lambda", "kappa", "delta", "EB", "white", "trend"), bounds = NULL,
  meserr = NULL, userstart = NULL, badLnL = 10000)
}
\arguments{
\item{phy}{Phylogenetic tree in phylo format}

\item{data}{data matrix with rownames equal to phy}

\item{data.names}{Tip names that match phy}

\item{model}{Choice of models include "BM", "OU", "lambda", "kappa",
"delta", "EB", "white", "trend"}

\item{bounds}{NULL for fitContinuous.hacked}

\item{meserr}{Measurement error, if any}

\item{userstart}{Optional user start input}

\item{badLnL}{Return value for bad likelihood calculations}
}
\value{
Returns a matrix of parameter estimates, along with approximate
standard errors and 95% confidence intervals.  Also returns the
log-likelihood of the model (lnl).
}
\description{
This is TreEvo's version of fitContinuous. This version differs from Geigers
in using a different optimization routine, lack of bounds, use of
pseudoinverse for matrix inversion, and multiple start values.
}
\details{
See Geiger's fitContinuous for details.
}
\examples{


data(simData)
fitContinuous.hacked(simPhy, simChar, model="lambda")

}
\references{
O'Meara and Banbury, unpublished; Harmon et al. 2007
}
\author{
Luke J. Harmon and Wendell Challenger, Brian O'Meara and Barb
Banbury
}
\keyword{fitContinuous.hacked}
