\name{getSimulationSplits}

\title{Simulation Splits}

\description{This function retreives branch lengths and ancestor-decendant relationships from a tree}

\usage{
getSimulationSplits(phy)
}

\arguments{
  \item{phy}{Tree (ape phylo format)}
}
\details{
	Used by other TreEvo functions
}
\value{
A data frame of branching times, ancestor and descendant vectors
}
\references{O'Meara and Banbury, unpublished}
\author{Brian O'Meara and Barb Banbury}
\seealso{
}
\examples{

}
\keyword{}

