% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highestPostDens.R
\name{highestPostDens}
\alias{highestPostDens}
\title{Highest Posterior Density}
\usage{
highestPostDens(particleDataFrame, percent = 0.95, returnData = F)
}
\arguments{
\item{particleDataFrame}{\code{particleDataFrame} output from \code{doRun}}

\item{percent}{Probability content of the highest posterior density (HPD).}

\item{returnData}{Option to return data that falls within HPD interval.}
}
\value{
Returns a matrix with weighted mean, standard deviation, upper and lower HPD for
each free parameter.
}
\description{
This function calculates the highest posterior density (HPD) for each freely varying parameter.
}
\examples{

data(simRun)

highestPostDens(results$particleDataFrame, percent=0.95, returnData=F)

}
\seealso{
\code{\link{HPDinterval}} in package \code{coda}
}
\author{
Brian O'Meara and Barb Banbury
}
