% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interparticleDistance.R
\name{interparticleDistance}
\alias{interparticleDistance}
\title{Calculate interparticle distance}
\usage{
interparticleDistance(x, y, abs = TRUE)
}
\arguments{
\item{x}{Vector of parameter values}

\item{y}{Vector of parameter values}

\item{abs}{If TRUE, uses absolute value for distance}
}
\value{
Returns a matrix of distances between pairwise points in the vectors
}
\description{
This function calculates interparticle distance between two vectors
}
\details{
IPD is used by several internal TreEvo function to compare two
distributions.
}
\references{
O'Meara and Banbury, unpublished
}
\author{
Brian O'Meara and Barb Banbury
}
