% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOptimaIntrinsic.R
\name{multiOptimaIntrinsic}
\alias{multiOptimaIntrinsic}
\title{Single Evolutionary Regime Model with Multiple Optima}
\usage{
multiOptimaIntrinsic(params, states, timefrompresent)
}
\arguments{
\item{params}{A vector containing input parameters for the given model (see \emph{Description} below on what parameters).}

\item{states}{Vector of current trait values for a taxon. May be multiple for some models, but generally expected to be
only a single value. Multivariate \code{TreEvo} is not yet supported.}

\item{timefrompresent}{The amount of time from the present - generally ignored except for time-dependent models.}
}
\value{
A vector of values representing character displacement of that lineage over a single time step.
}
\description{
This model describes an evolutionary process where multiple optima exist. Lineages are attracted
to an optima in their vicinity (this is handled as a stochastic process where which attract affects
which population at a given point in time is weighted with respect to the proximity of a given
population's trait value to the given optima), but as a lineage approaches the attractor that
controls it, it may experience less directional evolution in the direction of the optima,
and overall show more variation. This framework allows for the model to describe the evolution
with respect to multiple optima simultaneously, without needing to treat individual transitions
between optima (or macroevolutionary 'regimes') as a separate parameter. Thus, the optima exist
within a single evolutionary regime, with only a scaling parameter present that allows more
or less frequent transitions between optima by altering the impact of proximity to optima.
}
\details{
\code{multiOptimaIntrinsic} describes a model of intrinsic character evolution with
multiple regimes. New character values are generated after one time step via a discrete-time OU
process with a particular optima assigned to a particular regime, and each time-step 
a lineage has some finite probability of switching to a new regime, and being drawn to that regime's optima.
The chance of a lineage being drawn to a particular optima is based on proximity to
that optima, but the chance of switching to another regime is never completely negligible.
The strength of the draw to the optima, the attraction strength, 'alpha', is the same for all regimes (all optima).
This model has \emph{n} input parameters:
\code{multiOptimaIntrinsic} with \code{params = sigma} (rate of dispersion), 
\code{alpha} (strength of attraction to an optima),
\code{rho} (an exponent scaling the weighting of distance to optima --
this parameter will control the probability of a lineage switching optima),
and two or more \code{theta} values, which are the optima for the different macroevolutionary adaptiv regimes.
}
\examples{

# three optima model, with strong attraction	
set.seed(1)
params<-c(
	sigma=0.1,
	alpha=0.7,
	rho=1,
	theta=c(-20,20,50)
	)	
	
multiOptimaIntrinsic(params=params, states=0, timefrompresent=NA)

# simulate n time-steps, repeat many times, plot results
repeatSimSteps<-function(params,trait=0,nSteps){
	for(i in 1:nSteps){
	# add to original trait value to get new trait value
		trait<-trait+multiOptimaIntrinsic(
			params=params, states=trait, timefrompresent=NA)
		}
	trait
	}
repSim<-replicate(300,repeatSimSteps(params,trait=0,100))
hist(repSim,main="Simulated Trait Values",breaks=20)


# same model above, with more switching between optima
set.seed(1)
params<-c(
	sigma=0.1,
	alpha=0.7,
	rho=0.5,
	theta=c(-20,20,50)
	)	
	
multiOptimaIntrinsic(params=params, states=0, timefrompresent=NA)

# simulate n time-steps, repeat many times, plot results
repeatSimSteps<-function(params,trait=0,nSteps){
	for(i in 1:nSteps){
	# add to original trait value to get new trait value
		trait<-trait+multiOptimaIntrinsic(
			params=params, states=trait, timefrompresent=NA)
		}
	trait
	}
repSim<-replicate(300,repeatSimSteps(params,trait=0,100))
hist(repSim,main="Simulated Trait Values",breaks=20)

}
\seealso{
An alternative approach in \code{TreEvo} to estimating a macroevolutionary landscape with multiple optima
is the Fokker–Planck–Kolmogorov (FPK) model, which can be fit with \code{\link{landscapeFPK_Intrinsic}}.
This model does not assume a set number of optima nor that they have similar attractor strength, but
parameters may be difficult to interpret in isolation, and fitting this model may be slower with \code{TreEvo}
due to necessary linear algebra transformations. 
Other intrinsic models are described at \code{\link{intrinsicModels}}.
}
\author{
David W. Bapst
}
