\name{mutateState}
\alias{mutateState}

\title{Mutate Character State}


\usage{
mutateState(startingState, standardDevFactor, priorValues, priorFn)
}

\arguments{
  \item{startingState}{Character state prior to mutating}
  \item{standardDevFactor}{Standard deviation}
  \item{priorValues}{Vector of priorValues}
  \item{priorFn}{Prior shape}
}
\details{This function mutates the character state of a given taxon by one discrete time step
}
\value{
	}
\references{O'Meara and Banbury, unpublished}
\author{Brian O'Meara and Barb Banbury}

