% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abcmodels.extrinsic.R
\name{nullExtrinsic}
\alias{nullExtrinsic}
\title{Extrinsic Character Evolution Models}
\usage{
nullExtrinsic(params, selfstates, otherstates, timefrompresent)
}
\arguments{
\item{params}{describes input paramaters for the model}

\item{selfstates}{vector of states for each taxon}

\item{otherstates}{matrix of character states, one row per taxon and once
column per state}

\item{timefrompresent}{which time slice in the tree}
}
\value{
A matrix of values representing character displacement from a single
time step in the tree.
}
\description{
This function describes a model of no extrinsic character evolution
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}
\keyword{extrinsic}
\keyword{nullExtrinsic}

