% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pairings.R
\name{pairings}
\alias{pairings}
\title{Find unique pairings}
\usage{
pairings(nRuns)
}
\arguments{
\item{nRuns}{Numeric value representing haw many runs to compare}
}
\value{
Returns a matrix of combinations
}
\description{
This function finds all possible pairwise combinations for multiple runs.
}
\details{
Used in several internal functions for analysis of results
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}

