% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PairwiseESS.R
\name{PairwiseESS}
\alias{PairwiseESS}
\title{Pairwise Effective Sample Size}
\usage{
PairwiseESS(particleDataFrame)
}
\arguments{
\item{particleDataFrame}{particleDataFrame can be a single data frame or a
list of data frames}
}
\value{
Returns a matrix with ESS values of all pairwise runs
}
\description{
This function calculates ESS on results.  Performs the best when results are
from multiple runs.
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}

