% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseKS.R
\name{pairwiseKS}
\alias{pairwiseKS}
\title{Pairwise Kolmogorov-Smirnov test}
\usage{
pairwiseKS(particleDataFrameList)
}
\arguments{
\item{particleDataFrameList}{An object of type list, composed of particleDataFrames from separate analyses.}
}
\value{
Returns a matrix with Kolmogorov-Smirnov values of all pairwise runs.
}
\description{
This function calculates Kolmogorov-Smirnov on results.
}
\examples{

data(simRunExample)

pdfList<-list(Brownian=results$particleDataFrame,Bounded=resultsBound$particleDataFrame)

pairwiseKS(particleDataFrameList = pdfList)
}
\author{
Brian O'Meara and Barb Banbury
}
