% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotABC_3D.R
\name{plotABC_3D}
\alias{plotABC_3D}
\title{3D ABCplots}
\usage{
plotABC_3D(particleDataFrame, parameter, show.particles = "none",
  plot.parent = FALSE, realParam = FALSE, realParamValues = NA)
}
\arguments{
\item{particleDataFrame}{A \code{particleDataFrame} object, as found among the output from \code{\link{doRun}} functions.}

\item{parameter}{Column number of parameter of interest from
\code{particleDataFrame}.}

\item{show.particles}{Option to show particles on 3d plot as "none" or as a
function of "weights" or "distance".}

\item{plot.parent}{Option to plot lines on the floor of the 3d plot to show
particle parantage.}

\item{realParam}{Option to display real parameter value as a solid line, 
also must give actual value for this (realParamValues).  Note: this should
only be done with simulated data where real param values are recorded.}

\item{realParamValues}{Value for \code{realParam}.}
}
\description{
Plot posterior density distribution for each generation in 3d plot window
}
\details{
This opens a new interactive 3d plotting window and plots the posterior
density distribution of accepted particles from each generation.  Several
options are available to add to the plot: plotting particles by weight or
distance, plotting particle parantage, and plotting the real parameter
values (if known).

As of version 0.6.0, rejected particles are not saved for outputting by the parallelized algorithm, 
and thus they are no longer displayed by this function, unlike previous versions.
}
\note{
This function requires access functions \code{triangulate} and the \code{as} method for
class \code{gpc.poly} from package \code{gpclib}. As of 05-30-18, this package
was not available from CRAN as a Windows binary, and thus this function is likely
unavailable to many (if not all) Windows users.

This function also requires the package \code{rgl}, which is usually easier to
obtain than package \code{gpclib}, but may not be buildable on some UNIX workstations.
}
\examples{

# need to check for required suggested packages
if(requireNamespace("gpclib", quietly = TRUE) & requireNamespace("rgl", quietly = TRUE)){

 data(simRunExample)
 plotABC_3D(particleDataFrame = results[[1]]$particleDataFrame, parameter = 7, 
     show.particles = "none", plot.parent = FALSE, realParam = FALSE, realParamValues = NA)

}
}
\author{
Barb Banbury
}
