% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosteriors.R, R/plotPrior.R
\name{plotPosteriors}
\alias{plotPosteriors}
\alias{plotPosteriors}
\title{Plot posterior distributions}
\usage{
plotPosteriors(priorFn = match.arg(arg = priorFn, choices = c("fixed",
  "uniform", "normal", "lognormal", "gamma", "exponential"), several.ok =
  FALSE), priorVariables, plotQuants = TRUE, plotLegend = TRUE)

plotPosteriors(priorFn = match.arg(arg = priorFn, choices = c("fixed",
  "uniform", "normal", "lognormal", "gamma", "exponential"), several.ok =
  FALSE), priorVariables, plotQuants = TRUE, plotLegend = TRUE)
}
\arguments{
\item{priorFn}{Prior shape; either "fixed", "uniform", "normal",
"lognormal", "gamma", "exponential"}

\item{priorVariables}{Variables needed to describe the shape of the
distribution.  uniform distributions=c(min, max); normal
distributions=c(mean, sd); lognormal distributions=c(mean, sd); gamma
distributions=c(shape, scale); exponential distributions=c(rate)}

\item{plotQuants}{If TRUE, plots line segments at the quantiles}

\item{plotLegend}{If TRUE, plots legend box with quantile values}

\item{particleDataFrame}{particleDataFrame from TreEvo results, can be a
single data frame or a list of data frames}

\item{PriorMatrix}{PriorMatrix from TreEvo results, can be a single matrix
or a list of matrices}

\item{realParam}{Plot line segments where real parameter values are known}

\item{realParamValues}{Values for real paramters, include a value for each
parameter (inlcuding fixed values)}
}
\value{
Returns a plot for each free parameter.

Returns a distribution plot
}
\description{
This function plots posterior distributions from the last generation of each
free parameter

This function plots the distribution of priors
}
\details{
If the particleDataFrame is a list of seperate TreEvo runs, posteriors will
be layered over eachother to check for repeatability.  Grayscale of
posteriors depends on total number of runs.

If the PriorMatrix is a list of matrices, only the first matrix will be
plotted.  In other words, prior matrices MUST be identical for all runs.

realParam and realParamValues should only be used with simulated data, where
the true values are known.

This functions is to help visualize the shape of the priors for TreEvo abc
analyses.
}
\examples{

# data(Res)
# plotPosteriors(particleDataFrame=res$particleDataFrame,
#    PriorMatrix=res$priorMatrix, realParam=FALSE, realParamValues=NA)


	plotPrior(priorFn="exponential", priorVariables=c(10))
	
	plotPrior("normal", c(1,2))
	
	plotPrior("gamma", c(2, .2), plotQuants=FALSE, plotLegend=FALSE)

}
\author{
Barb Banbury and Brian O'Meara

Brian O'Meara and Barb Banbury
}
