% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPriorPost.R
\name{plotPriorPost}
\alias{plotPriorPost}
\alias{plotPrior}
\alias{getUnivariatePriorCurve}
\alias{getUnivariatePosteriorCurve}
\alias{plotUnivariatePosteriorVsPrior}
\title{Plotting, Sampling and Comparing the Prior and Posterior Distributions}
\usage{
plotPrior(priorFn = match.arg(arg = priorFn, choices = c("fixed", "uniform",
  "normal", "lognormal", "gamma", "exponential"), several.ok = FALSE),
  priorVariables, plotQuants = TRUE, plotLegend = TRUE)

getUnivariatePriorCurve(priorValues, priorFn, nPoints = 1e+05, from = NULL,
  to = NULL, prob = 0.95)

getUnivariatePosteriorCurve(acceptedValues, from = NULL, to = NULL,
  prob = 0.95)

plotUnivariatePosteriorVsPrior(posteriorCurve, priorCurve,
  label = "parameter", trueValue = NULL, prob = 0.95)
}
\arguments{
\item{priorFn}{Prior Shape of the distribution; one of either "fixed", "uniform", "normal",
"lognormal", "gamma", or "exponential".}

\item{priorVariables}{Variables needed to describe the shape of the
distribution, dependent on \code{priorFn}: 
\item{priorFn = "uniform"}{priorVariables = c(min, max)} 
\item{priorFn = "normal"}{priorVariables = c(mean, standard deviation)}
\item{priorFn = "lognormal"}{priorVariables = c(mean, standard deviation)}
\item{priorFn = "gamma"}{priorVariables = c(shape, scale)}
\item{priorFn = "exponential"}{priorVariables = c(rate)}}

\item{plotQuants}{If \code{TRUE}, plots line segments at the quantiles}

\item{plotLegend}{If \code{TRUE}, plots legend box with quantile values}

\item{nPoints}{Number of points to draw.}

\item{from}{Lower bound, if any.}

\item{to}{Upper bound, if any.}

\item{prob}{Probability content of the highest posterior density (HPD).}

\item{acceptedValues}{Vector of accepted particle values}

\item{posteriorCurve}{Kernal density estimates for the posterior}

\item{priorCurve}{Kernal density estimates for the prior}

\item{label}{X label for plot}

\item{trueValue}{True parameter value, if any}

\item{prob}{Probability content of HPD}
}
\value{
\code{plotPrior} and \code{plotUnivariatePosteriorVsPrior} produce plots of the respective distributions.

\code{getUnivariatePriorCurve} and \code{getUnivariatePosteriorCurve} returns a list of x and y density coordinates, mean, and lower and
upper highest posterior density (HPD), for their respective distribution.
}
\description{
Assorted functions for visualizing the prior and posterior probability distributions associated with ABC analyses.
}
\details{
Function \code{plotPrior} visualizes the shape of various prior probability distributions available in TreEvo ABC analyses.

Function \code{getUnivariatePriorCurve} pulls random values from a selected prior probability distribution.

Function \code{getUnivariatePosteriorCurve} pulls values from the posterior distribution.

Function \code{plotUnivariatePosteriorVsPrior} plots the univariate density distributions from the prior and posterior.
}
\examples{

	plotPrior(priorFn="exponential", priorVariables=c(10))
	
	plotPrior(priorFn="normal", priorVariables=c(1,2))
	
	plotPrior(priorFn="gamma", priorVariables=c(2, .2), plotQuants=FALSE, plotLegend=FALSE)

}
\author{
Brian O'Meara and Barb Banbury
}
