% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSTransform.R
\name{PLSTransform}
\alias{PLSTransform}
\title{Partial Least Squares Transform}
\usage{
PLSTransform(summaryValuesMatrix, pls.model)
}
\arguments{
\item{summaryValuesMatrix}{Matrix of summary statistics from simulations}

\item{pls.model}{Results output by from \code{\link{returnPLSModel}}}
}
\value{
Returns transformed summary statistics
}
\description{
This function uses results from a Partial Least Squares (PLS) model to transform summary values.
}
\examples{

#PLSTransform(summaryValuesMatrix, pls.model)

}
\author{
Brian O'Meara and Barb Banbury
}
