% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRun.R
\docType{data}
\name{simRun}
\alias{simRun}
\alias{simPhy}
\alias{simChar}
\alias{results}
\title{Simulated data}
\format{simPhy in phylo format and simChar in single column matrix and results}
\description{
Simulated 30-taxon coalescent tree and simulated character data from BM
intrinsic model. Character data was generated using doSimulation.
Also includes results from an example analysis.
}
