% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatsLong.R
\name{solnfreq}
\alias{solnfreq}
\title{Calculate frequency of finding best solution in Geiger}
\usage{
solnfreq(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A returned object from fitContinuous()}

\item{tol}{Tolerance for equality of solutions}
}
\value{
The frequency with which the best solution was found
}
\description{
This function is taken from an internal Geiger function
}
