% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePosterior.R
\name{summarizePosterior}
\alias{summarizePosterior}
\title{Summarize Posterior Distribution for a Free Parameter}
\usage{
summarizePosterior(particleDataFrame, alpha = 0.8, coda = FALSE,
  verboseMultimodal = TRUE, ...)
}
\arguments{
\item{particleDataFrame}{A \code{particleDataFrame} object, as
found among the output from \code{\link{doRun}} functions.}

\item{alpha}{The threshold used for defining the highest density frequency
cut-off. If the highest density interval is applied to a Bayesian MCMC
posterior sample, then the interval is effectively calculated for this
value as a posterior probability density.}

\item{coda}{Default is \code{FALSE}. If \code{TRUE}, unimodal highest density regions will
instead be calculated using \code{\link{HPDinterval}} from package \code{coda}, which is
similar to the function \code{quantile} in that it calculates only a single interval.}

\item{verboseMultimodal}{If \code{TRUE}, the function will print a message
indicating when the inferred highest density interval is discontinuous, and
thus likely reflects that the supplied data is multimodal.}

\item{...}{Additional arguments passed to \code{\link{density}}. 
A user may want to mess with this to adjust bandwidth, et cetera.}
}
\value{
Returns a list, wherein each element of the list is secondary list containing
the weighted mean, standard deviation, and a matrix giving the highest density
intervals (e.g. the highest posterior density intervals). Because posterior
estimates of parameter values may be multimodal, multiple sets of bounds
may be reported for complex posterior distributions, which each constitute one 
row of the output matrix. See \code{\link{highestDensityInterval}} for details.
}
\description{
This function summarizes the distribution of parameter estimates from
the posterior of an ABC analysis sing the \code{doRun} functions in \code{TreEvo}, for all
freely varying parameters. Only the final generation is considered. This summary includes
the mean, standard deviation and Highest Posterior Density (at a 0.8 alpha) for each parameter.
}
\examples{

data(simRunExample)

summarizePosterior(results[[1]]$particleDataFrame, alpha = 0.5)

}
\seealso{
This function is essentially a wrapper for independently applying 
a few summary statistics and applying
\code{\link{highestDensityInterval}} to multiple parameter estimates, taken from
the last generation of an ABC analysis in \code{TreEvo}. As each parameter
is handled independently, the returned HPD intervals may not properly account
for covariation among parameter estimates from the posterior. If testing
whether a given observation is within a given density of the posterior or
not, please look at function \code{\link{testMultivarOutlierHDR}}.
}
\author{
David W Bapst
}
