% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTaxonStates.R
\name{summarizeTaxonStates}
\alias{summarizeTaxonStates}
\title{Create a data frame of taxon states}
\usage{
summarizeTaxonStates(taxa)
}
\arguments{
\item{taxa}{a list of objects}
}
\value{
Returns a data frame of taxon values
}
\description{
This function creates a data frame of taxon states while simulating
characters with doSimulation and doSimulationsForPlotting TreEvo functions
}
\details{
Used by TreEvo doSimulation and doSimulationForPlotting functions to
summarize a list of objects into a data frame of taxon values
}
\author{
Brian O'Meara and Barb Banbury
}
