% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryStatsLong.R
\name{summaryStatsLong}
\alias{summaryStatsLong}
\title{Calculate summary statistics}
\usage{
summaryStatsLong(phy, traits)
}
\arguments{
\item{phy}{Tree (Phylogenetic tree in phylo format)}

\item{traits}{data matrix with rownames equal to phy}
}
\value{
Returns a vector of summary statistics
}
\description{
This function creates a vector of summary statistics for TreEvo analysis
}
\details{
Caluclates 17 summary statistics from the fitContinuous.hacked function
(brown.lnl, brown.beta, brown.aic, lambda.lnl, lambda.beta, lambda.lambda,
lambda.aic, delta.lnl, delta.beta, delta.delta, delta.aic, ou.lnl, ou.beta,
ou.alpha, ou.aic, white.lnl, white.aic), plus raw.mean, raw.max, raw.min,
raw.var, raw.median, and all tip character values, phylogenetic independent
contrasts, ancestral state reconstruction values, and the range of ancestral
state reconstruction confidence interval.
}
\examples{
#summaryStatsLong(phy, char)
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}
\keyword{summaryStatsLong}

