\name{sumStatNames}
\alias{sumStatNames}

\title{Naming function for summary stats}

\description{
	This function creates a vector of summary stat names}
\usage{
sumStatNames(phy)
}

\arguments{
  \item{phy}{Tree (Phylogenetic tree in phylo format)}
}

\references{O'Meara and Banbury, unpublished}
\author{Brian O'Meara and Barb Banbury}

