% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treevo-package.R
\docType{package}
\name{TreEvo-package}
\alias{TreEvo-package}
\alias{treevo}
\title{TreEvo--abc for comparative methods}
\description{
A package for applying Approximate Bayesian Computation to estimating parameters of trait evolution in comparative analyses.
}
\details{
\tabular{ll}{ Package: \tab TreEvo\cr Type: \tab Package\cr Version: \tab
0.3.3\cr Date: \tab 2012-07-02\cr License: \tab GPL\cr }
}
\examples{

data(simRun)

doRun_prc(
  phy = simPhy,
  traits = simChar,
  intrinsicFn=brownianIntrinsic,
  extrinsicFn=nullExtrinsic,
  startingPriorsFns="normal",
  startingPriorsValue=matrix(c(mean(simChar[,1]), sd(simChar[,1]))),
  intrinsicPriorsFns=c("exponential"),
  intrinsicPriorsValues=matrix(c(10, 10), nrow=2, byrow=FALSE),
  extrinsicPriorsFns=c("fixed"),
  extrinsicPriorsValues=matrix(c(0, 0), nrow=2, byrow=FALSE),
  TreeYears=1000,
  standardDevFactor=0.2,
  plot=FALSE,
  StartSims=300,
  epsilonProportion=0.7,
  epsilonMultiplier=0.7,
  nStepsPRC=5,
  numParticles=100,
  jobName="examplerun_prc",
  stopRule=FALSE,
  multicore=FALSE,
  coreLimit=1
)


}
\author{
Brian O'Meara, Barb L. Banbury, David W. Bapst

Maintainer: David Bapst <dwbapst@gmail.com>
}
