% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abctaxon.R
\name{abctaxon}
\alias{abctaxon}
\title{abctaxon}
\usage{
abctaxon(id = 0, name = "taxon0", timeSinceSpeciation = 0, states = NA,
  nextstates = NA)
}
\arguments{
\item{id}{Simulation ID}

\item{name}{Taxon of interest}

\item{timeSinceSpeciation}{Time since speciation}

\item{states}{States of taxon}

\item{nextstates}{States of taxon after mutation}
}
\description{
Internal TreEvo function to create a list of objects
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}

