% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abctaxon.R
\name{abctaxon}
\alias{abctaxon}
\title{abctaxon}
\usage{
abctaxon(id = 0, name = "taxon0", timeSinceSpeciation = 0, states = NA,
  nextstates = NA, startTime = 0, endTime = Inf, terminal = TRUE,
  ancestorId = NA)
}
\arguments{
\item{id}{Simulation ID}

\item{name}{Taxon of interest}

\item{timeSinceSpeciation}{Time since speciation}

\item{states}{States of taxon}

\item{nextstates}{States of taxon after mutation}

\item{startTime}{When this lineage starts}

\item{endTime}{When this lineage stops}

\item{terminal}{If TRUE, is a terminal taxon}

\item{ancestorId}{ID of the ancestor of this node.}
}
\description{
Internal TreEvo function to create a list of objects
}
\references{
O'Meara and Banbury, unpublished
}
\author{
Brian O'Meara and Barb Banbury
}
