% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesCoverageProb.R
\name{bayesCoverageProb}
\alias{bayesCoverageProb}
\title{Bayesian Coverage Probability}
\usage{
bayesCoverageProb(RealParam, HPD, verbose = F)
}
\arguments{
\item{RealParam}{Real parameter values.}

\item{HPD}{List of highest posterior density from \code{doRun_rej} or \code{doRun_prc} results}

\item{verbose}{If \code{TRUE}, commented screen output is produced.}
}
\value{
Returns a value for each free parameter that describes the
percentage that the real value falls within the HPD.
}
\description{
This function calculates coverage probability for a list of highest posterior densities (HPDs) for a set of parameters.
}
\details{
Only for use with simulated data to test models.
}
\examples{

breakThisExample
}
\author{
Brian O'Meara and Barb Banbury
}
