% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesCoverageProb.R
\name{bayesCoverageProb}
\alias{bayesCoverageProb}
\title{Bayesian Coverage Probability}
\usage{
bayesCoverageProb(RealParam, HPD, verbose = FALSE)
}
\arguments{
\item{RealParam}{Real parameter values.}

\item{HPD}{List of highest posterior density from \code{\link{doRun}} results.}

\item{verbose}{If \code{TRUE}, commented screen output is produced.}
}
\value{
Returns a value for each free parameter that describes the
percentage that the real value falls within the HPD.
}
\description{
This function calculates coverage probability for a list of highest posterior densities (HPDs) for a set of parameters.
}
\details{
Only for use with simulated data to test models.
}
\examples{

data(simRun)

# real (generating) parameters
genPar<-c(ancState,genRate)

HPDs<-list(results$HPD,resultsBound$HPD)

bayesCoverageProb(RealParam=genPar, HPD=HPDs, verbose=TRUE)

}
\author{
Brian O'Meara and Barb Banbury
}
