% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcoxEstimation.R
\name{boxcoxEstimation}
\alias{boxcoxEstimation}
\title{Box-Cox Estimation}
\usage{
boxcoxEstimation(summaryValuesMatrix)
}
\arguments{
\item{summaryValuesMatrix}{Matrix of summary statistics from simulations}
}
\value{
Returns a matrix of Box-Cox transformed summary statistics with the
same dimensions as summaryValuesMatrix.
}
\description{
This function Box-Cox transforms summary values
}
\examples{

#data(res)

#boxcoxEstimation(summaryValuesMatrix)

}
\author{
Brian O'Meara and Barb Banbury
}
