% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcoxTransformation.R
\name{boxcoxTransformation}
\alias{boxcoxTransformation}
\alias{boxcoxTransformationMatrix}
\title{Box-Cox Transformation for Simulation Summary Values}
\usage{
boxcoxTransformation(summaryValuesVector, boxcoxAddition, boxcoxLambda)

boxcoxTransformationMatrix(summaryValuesMatrix)
}
\arguments{
\item{summaryValuesVector}{Vector of summary statistics from a single simulation}

\item{boxcoxAddition}{Vector of boxcox additions from boxcoxEstimation}

\item{boxcoxLambda}{Vector of boxcox lambda values from boxcoxEstimation}

\item{summaryValuesMatrix}{Matrix of summary statistics from simulations}
}
\value{
\code{boxcoxTransformation} returns a vector of Box-Cox transformed summary statistics from a
single observation. \code{boxcoxTransformationMatrix} returns a matrix of Box-Cox transformed summary statistics with the
same dimensions as \code{summaryValuesMatrix}.
}
\description{
\code{boxcoxTransformation} Box-Cox transforms summary values from a single simulation, while
\code{boxcoxTransformationMatrix} Box-Cox transforms summary values from multiple simulations.
}
\examples{

data(simRun)

boxcoxTransformationMatrix(summaryValuesMatrix=results$summaryValuesMatrix)

boxcoxTransformation(summaryValuesVector=results$summaryValuesMatrix, boxcoxAddition, boxcoxLambda)

}
\author{
Brian O'Meara and Barb Banbury
}
