% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxcoxTransformation.R
\name{boxcoxTransformation}
\alias{boxcoxTransformation}
\title{Box-Cox Transformation}
\usage{
boxcoxTransformation(summaryValues, boxcoxAddition, boxcoxLambda)
}
\arguments{
\item{summaryValues}{Vector of summary statistics}

\item{boxcoxAddition}{Vector of boxcox additions from boxcoxEstimation}

\item{boxcoxLambda}{Vector of boxcox lambda values from boxcoxEstimation}
}
\value{
Returns a vector of Box-Cox transformed summary statistics from a
single observation.
}
\description{
This function Box-Cox transforms summary values from a single simulation
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished; Bates et al. 2009
}
\keyword{Box-Cox}
\keyword{boxcoxTransformation}

