\name{compareListIPD}
\alias{compareListIPD}
\title{Plot interparticle distance between runs}
\description{
	This function plots interparticle distance between runs for each free parameter     
}
\usage{compareListIPD(particleDataFrame, verbose=F)}
\arguments{
  \item{particleDataFrame}{particleDataFrame output from doRun}
  \item{verbose}{Commented screen output}
}
\value{Returns a plot with IPD between runs per generation}

\references{O'Meara and Banbury, unpublished}
\author{Brian O'Meara and Barb Banbury}
