% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareListIPD.R
\name{compareListIPD}
\alias{compareListIPD}
\title{Plot interparticle distance between runs}
\usage{
compareListIPD(particleDataFrame, verbose = F)
}
\arguments{
\item{particleDataFrame}{particleDataFrame output from doRun}

\item{verbose}{Commented screen output}
}
\value{
Returns a plot with IPD between runs per generation
}
\description{
This function plots interparticle distance between runs for each free
parameter
}
\author{
Brian O'Meara and Barb Banbury
}
