% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareListIPD.R
\name{compareListIPD}
\alias{compareListIPD}
\title{Plot Interparticle Distance between Runs}
\usage{
compareListIPD(particleDataFrameList, verbose = FALSE)
}
\arguments{
\item{particleDataFrameList}{A list composed of multiple \code{particleDataFrame} objects, as output by \code{\link{doRun}} functions.}

\item{verbose}{Commented screen output.}
}
\value{
Produces a plot with IPD between runs per generation.
}
\description{
This function plots interparticle distance (IPD) between runs for each free
parameter.
}
\examples{

data(simRunExample)

pdfList <- list(results[[1]]$particleDataFrame, resultsBound[[1]]$particleDataFrame)

compareListIPD(particleDataFrameList = pdfList, verbose = TRUE)

}
\author{
Brian O'Meara and Barb Banbury
}
