\name{convertTaxonFrameToGeigerData}

\title{Converting Simulated Data to Geiger Formatted Data}
\description{
	This function takes simulated data and converts it to geiger's formatting.
}
\usage{
convertTaxonFrameToGeigerData(taxonframe, phy)
}
\arguments{
  \item{taxonframe}{data frame with species values}
  \item{phy}{Tree (ape phylo format)}
}
\details{
}
\value{
Returns a data frame in geiger format

	}
\references{O'Meara and Banbury, unpublished}
\author{Brian O'Meara and Barb Banbury}
\seealso{
}
\examples{

}
\keyword{}

