% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CredInt.R
\name{CredInt}
\alias{CredInt}
\title{Bayesian Credible Interval}
\usage{
CredInt(particleDataFrame, percent = 0.95)
}
\arguments{
\item{particleDataFrame}{particleDataFrame output from doRun}

\item{percent}{Probability content of HPD}
}
\value{
Returns a matrix with weighted mean, sd, upper and lower credible
intervals for each free parameter
}
\description{
This function calculates credible interval for each free parameter
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}

