% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSimulation.R
\name{doSimulation}
\alias{doSimulation}
\alias{doSimulationForPlotting}
\alias{doSimulationWithPossibleExtinction}
\title{Discrete Time Character Simulation}
\usage{
doSimulation(phy = NULL, taxon.df = NULL, intrinsicFn, extrinsicFn,
  startingValues, intrinsicValues, extrinsicValues, timeStep,
  saveHistory = FALSE, saveRealParams = FALSE, jobName = "")

doSimulationForPlotting(phy = NULL, taxon.df = NULL, intrinsicFn,
  extrinsicFn, startingValues, intrinsicValues, extrinsicValues, timeStep,
  plot = FALSE, savePlot = FALSE, saveHistory = FALSE,
  saveRealParams = FALSE, jobName = "")

doSimulationWithPossibleExtinction(phy = NULL, taxon.df = NULL, intrinsicFn,
  extrinsicFn, startingValues, intrinsicValues, extrinsicValues, timeStep,
  saveHistory = FALSE, saveRealParams = FALSE, jobName = "",
  returnAll = FALSE, verbose = FALSE, reject.NaN = TRUE)
}
\arguments{
\item{phy}{A phylogenetic tree, in package \code{ape}'s \code{phylo} format.}

\item{intrinsicFn}{Name of (previously-defined) function that governs how
traits evolve within a lineage, regardless of the trait values of other taxa.}

\item{extrinsicFn}{Name of (previously-defined) function that governs how
traits evolve within a lineage, based on their own ('internal') trait vlaue and the trait values
of other taxa.}

\item{startingValues}{State at the root.}

\item{intrinsicValues}{Vector of values corresponding to the parameters of the
intrinsic model.}

\item{extrinsicValues}{Vector of values corresponding to the parameters of the
extrinsic model.}

\item{timeStep}{This value corresponds to the number of discrete time steps
on the shortest branch}

\item{saveHistory}{If \code{TRUE}, saves the character history throughout the simulation}

\item{saveRealParams}{Saves \code{intrinsicValues} and \code{extrinsicValues} as both a
matrix and a vector to an external .Rdata file.}

\item{jobName}{Optional name for the job.}

\item{plot}{Will create a new interactive window that plots character values
throughout the history of the tree.}

\item{savePlot}{Saves the character tree using \code{jobName}.}

\item{returnAll}{If \code{TRUE}, the output returned is a \code{data.frame} containing the values at each node from the simulation.}

\item{verbose}{If \code{TRUE}, gives messages about how the simulation is progessing via \code{print}.}

\item{reject.NaN}{If \code{TRUE}, stop run if any simulated value is NaN}
}
\value{
A data frame of species character (tip) values in the tree
(unless \code{returnAll = TRUE}, in which case the raw \code{data.frame} from the simulation).
}
\description{
These simulation function evolve continuous characters under a discrete time process.
}
\details{
When \code{saveHistory} is \code{TRUE}, processor time will increase quite a bit.
\code{SaveRealParams} is useful for tracking the "real" true values if simulating
data for abc runs.  It is not useful for empirical abc runs.
}
\examples{

tree<-rcoal(30)

#Simple Brownian motion
char<-doSimulation(
	phy=tree,
	intrinsicFn=brownianIntrinsic,
	extrinsicFn=nullExtrinsic,
	startingValues=c(10), #root state
	intrinsicValues=c(0.01),
	extrinsicValues=c(0),
	timeStep=0.0001,
	saveHistory=FALSE)


#Character displacement model with minimum bound
char<-doSimulation(
	phy=tree,
	intrinsicFn=boundaryMinIntrinsic,
	extrinsicFn=ExponentiallyDecayingPushExtrinsic,
	startingValues=c(10), #root state
	intrinsicValues=c(0.05, 10, 0.01),
	extrinsicValues=c(0, .1, .25),
	timeStep=0.001,
	saveHistory=FALSE)

#Simple Brownian motion
char<-doSimulationForPlotting(
	phy=tree,
	intrinsicFn=brownianIntrinsic,
	extrinsicFn=nullExtrinsic,
	startingValues=c(10), #root state
	intrinsicValues=c(0.01),
	extrinsicValues=c(0),
	timeStep=0.0001,
	plot=FALSE,
	saveHistory=FALSE)


#Character displacement model with minimum bound
char<-doSimulationForPlotting(
	phy=tree,
	intrinsicFn=boundaryMinIntrinsic,
	extrinsicFn=ExponentiallyDecayingPushExtrinsic,
	startingValues=c(10), #root state
	intrinsicValues=c(0.05, 10, 0.01),
	extrinsicValues=c(0, .1, .25),
	timeStep=0.001,
	plot=TRUE,
	saveHistory=FALSE)


tree<-rcoal(30)

#Simple Brownian motion
char<-doSimulationWithPossibleExtinction(
	phy=tree,
	intrinsicFn=brownianIntrinsic,
	extrinsicFn=nullExtrinsic,
	startingValues=c(10), #root state
	intrinsicValues=c(0.01),
	extrinsicValues=c(0),
	timeStep=0.0001,
	saveHistory=FALSE)

}
\author{
Brian O'Meara and Barb Banbury

Brian O'Meara and Barb Banbury
}
