% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBMRatePrior.R
\name{GetBMRatePrior}
\alias{GetBMRatePrior}
\title{Get BM Rate Prior}
\usage{
GetBMRatePrior(phy, traits, timeStep)
}
\arguments{
\item{phy}{Tree (Phylogenetic tree in phylo format)}

\item{traits}{data matrix with rownames equal to phy}

\item{timeStep}{time in a single iteration of the discrete-time simulation}
}
\value{
Returns a matrix of prior values
}
\description{
This function automatically calculates prior distributions for BM model of
evolution
}
\details{
Returns a matrix of prior values that can be used in the doRun functions.
Builds on functions in geiger to estimate distribution.
}
\examples{

#GetBMRatePrior(phy, traits, timeStep)


}
\references{
O'Meara and Banbury, unpublished
}
\author{
Brian O'Meara and Barb Banbury
}
\keyword{GetBMRatePrior}
