% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBMRatePrior.R
\name{getBMRatePrior}
\alias{getBMRatePrior}
\title{Get Brownian Motion Rate Prior}
\usage{
getBMRatePrior(phy, traits, timeStep, verbose = TRUE)
}
\arguments{
\item{phy}{A phylogenetic tree, in package \code{ape}'s \code{phylo} format.}

\item{traits}{Data matrix with rownames identical to \code{phy@tip.label}.}

\item{timeStep}{time in a single iteration of the discrete-time simulation}

\item{verbose}{If \code{TRUE}, gives messages about how the simulation is progessing via \code{print}.}
}
\value{
Returns a matrix of prior values
}
\description{
This function automatically calculates prior distributions for
the rate of trait evolution under the Brownian Motion (BM) model.
}
\details{
Returns a matrix of prior values that can be used in the \code{doRun} functions.
Builds on functions in \code{phylolm} to estimate distribution.
}
\examples{

data(simRunExample)

getBMRatePrior(phy=simPhy, traits=simChar, timeStep=1)


}
\author{
Brian O'Meara and Barb Banbury
}
