% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimulationSplits.R
\name{getSimulationSplits}
\alias{getSimulationSplits}
\title{Simulation Splits}
\usage{
getSimulationSplits(phy)
}
\arguments{
\item{phy}{Tree (ape phylo format)}
}
\value{
A data frame of branching times, ancestor and descendant vectors
}
\description{
This function retrieves branch lengths and ancestor-decendant relationships
from a tree
}
\details{
This function is used by other TreEvo functions for internal calculations.
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}

