% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonDFWithPossibleExtinction.R
\name{getTaxonDFWithPossibleExtinction}
\alias{getTaxonDFWithPossibleExtinction}
\title{Get data.frame of taxa}
\usage{
getTaxonDFWithPossibleExtinction(phy)
}
\arguments{
\item{phy}{Tree (ape phylo format)}
}
\value{
A list of abctaxon objects
}
\description{
This function retrieves branch lengths and ancestor-decendant relationships
from a tree
}
\details{
This function is used by other TreEvo functions for internal calculations.
}
\references{
O'Meara and Banbury, unpublished
}
\author{
Brian O'Meara
}
