% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotUnivariatePosteriorVsPrior.R
\name{getUnivariatePriorCurve}
\alias{getUnivariatePriorCurve}
\title{getUnivariatePriorCurve}
\usage{
getUnivariatePriorCurve(priorValues, priorFn, nPoints = 1e+05, from = NULL,
  to = NULL, prob = 0.95)
}
\arguments{
\item{priorValues}{Variables needed to describe the shape of the
distribution.  uniform distributions=c(min, max); normal
distributions=c(mean, sd); lognormal distributions=c(mean, sd); gamma
distributions=c(shape, scale); exponential distributions=c(rate)}

\item{priorFn}{Prior shape; either "fixed", "uniform", "normal",
"lognormal", "gamma", "exponential"}

\item{nPoints}{Number of points}

\item{from}{Lower bound, if any}

\item{to}{Upper bound, if any}

\item{prob}{Probability content of HPD}
}
\value{
Returns a list of x and y density coordinates, mean, and lower and
upper HPD
}
\description{
This function pulls random values from the prior
}
\author{
Barb Banbury and Brian O'Meara
}
\references{
O'Meara and Banbury, unpublished
}
\keyword{getUnivariatePriorCurve}

