% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutateState.R
\name{mutateState}
\alias{mutateState}
\title{Mutate Character State}
\usage{
mutateState(startingState, standardDevFactor, priorValues, priorFn)
}
\arguments{
\item{startingState}{Character state prior to mutating}

\item{standardDevFactor}{Standard deviation}

\item{priorValues}{Vector of priorValues}

\item{priorFn}{Prior shape}
}
\description{
This function mutates the character state of a given taxon by one discrete
time step
}
\examples{

breakThisExample
}
\author{
Brian O'Meara and Barb Banbury
}
