\name{pairings}
\alias{pairings}
\title{Find unique pairings}
\description{
	This function finds all possible pairwise combinations for multiple runs.  
}
\usage{pairings(nRuns)}
\arguments{
  \item{nRuns}{Numeric value representing haw many runs to compare}
}
\value{
	Returns a matrix of combinations
}
\details{Used in several internal functions for analysis of results}
\references{O'Meara and Banbury, unpublished}
\author{Brian O'Meara and Barb Banbury}
