% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseESS.R
\name{pairwiseESS}
\alias{pairwiseESS}
\title{Pairwise Effective Sample Size}
\usage{
pairwiseESS(inputData)
}
\arguments{
\item{inputData}{inputData can be a single data frame or a
list of data frames}
}
\value{
Returns a matrix with ESS values of all pairwise runs
}
\description{
This function calculates Effective Sample Size (ESS) on results.  Performs the best when results are
from multiple runs.
}
\examples{

data(simRunExample)

# this will give a warning
pairwiseESS(results[[1]]$particleDataFrame)

# ESS should be calculated over multiple runs
pairwiseESS(results)

# you can also manually assemble a list of particleDataFrame tables
    # and use this as the input
inputList<-list(results[[2]]$particleDataFrame,results[[1]]$particleDataFrame)
pairwiseESS(inputList)
}
\author{
Brian O'Meara and Barb Banbury
}
