% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairwiseESS.R
\name{pairwiseESS}
\alias{pairwiseESS}
\title{Pairwise Effective Sample Size}
\usage{
pairwiseESS(particleDataFrame)
}
\arguments{
\item{particleDataFrame}{particleDataFrame can be a single data frame or a
list of data frames}
}
\value{
Returns a matrix with ESS values of all pairwise runs
}
\description{
This function calculates Effective Sample Size (ESS) on results.  Performs the best when results are
from multiple runs.
}
\examples{

data(simRunExample)

pairwiseESS(results$particleDataFrame)
}
\author{
Brian O'Meara and Barb Banbury
}
