% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairwiseESS.R
\name{pairwiseESS}
\alias{pairwiseESS}
\title{Pairwise Effective Sample Size}
\usage{
pairwiseESS(inputData)
}
\arguments{
\item{inputData}{The input dataset can be a single data frame or a
list of data frames.}
}
\value{
Returns a matrix with ESS values of all pairwise runs.
}
\description{
This function calculates Effective Sample Size (ESS) on results.  
The ESS algorithm performs best when results are
from multiple runs.
}
\examples{

data(simRunExample)

# this will give a warning
pairwiseESS(resultsBMExample[[1]]$particleDataFrame)

# ESS should be calculated over multiple runs
pairwiseESS(resultsBMExample)

# you can also manually assemble a list of particleDataFrame tables
    # and use this as the input
inputList <- list(
    resultsBMExample[[2]]$particleDataFrame, 
    resultsBMExample[[1]]$particleDataFrame
    )
pairwiseESS(inputList)
}
\author{
Brian O'Meara and Barb Banbury
}
