% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairwiseKS.R
\name{PairwiseKS}
\alias{PairwiseKS}
\title{Pairwise Kolmogorov-Smirnov test}
\usage{
PairwiseKS(particleDataFrame)
}
\arguments{
\item{particleDataFrame}{List of particleDataFrames from separate runs}
}
\value{
Returns a matrix with Kolmogorov-Smirnov values of all pairwise runs
}
\description{
This function calculates Kolmogorov-Smirnov on results.
}
\examples{

data(simRun)

PairwiseKS(particleDataFrame)
}
\author{
Brian O'Meara and Barb Banbury
}
