% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentOffspringPlots.R
\name{parentOffspringPlots}
\alias{parentOffspringPlots}
\title{Plotting Parent-Offspring Particles}
\usage{
parentOffspringPlots(particleDataFrame)
}
\arguments{
\item{particleDataFrame}{\code{particleDataFrame} output by TreEvo ABC analyses.}
}
\value{
Creates a set of parent-offspring plots.
}
\description{
This function uses the \code{particleDataFrame} output by TreEvo ABC analyses and plots
parent-offspring particles from generation to generation.
}
\details{
Each parameter is plotted twice for parent-offspring relationships through
the generations.  In the top row, particles are plotted as a measure of
distance to the observed data; the farther away the particle the bigger the
circle.  Gray circles indicate rejected particles.  In the bottom row,
particles are plotted as a measure of their weights; larger circles are
closer to observed data and therefore carry more weight in the analysis.
Grayed out stars represent rejected particles.
}
\examples{

data(simRunExample)
parentOffspringPlots(results$particleDataFrame)

}
\author{
Brian O'Meara and Barb Banbury
}
