% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosteriors.R
\name{plotPosteriors}
\alias{plotPosteriors}
\title{Plot posterior distributions}
\usage{
plotPosteriors(particleDataFrame, priorsMat, realParam = FALSE,
  realParamValues = NA)
}
\arguments{
\item{particleDataFrame}{A \code{particleDataFrame} object returned by TreEvo ABC analyses, can be a
single data frame or a list of data frames.}

\item{realParam}{If \code{TRUE}, this function will plot line segments where real
parameter values are known. (Usually only true when simulated data is analyzed.)}

\item{realParamValues}{Values for real paramters, include a value for each
parameter (inlcuding fixed values). Otherwise should be \code{NA}.}

\item{PriorMatrix}{A \code{PriorMatrix} object returned by TreEvo ABC analyses, can be a single matrix
or a list of matrices. If a list, only the first matrix is used, all others are discarded.}
}
\value{
Returns a plot for each free parameter.
}
\description{
This function plots posterior distributions from the last generation of each
free parameter.
}
\details{
If the \code{particleDataFrame} is a list of seperate TreEvo runs, posteriors will
be layered over each other to check for repeatability.  The relative grayscale of
posterior distributions in the plot depends on their total number of runs.

If the \code{PriorMatrix} is a list of matrices, only the first matrix will be
plotted and others will be ignores.  
In other words, prior matrices MUST be identical for all runs.

\code{realParam} and \code{realParamValues} should only be used with simulated data, where
the true values are known.
}
\examples{

data(simRun)

plotPosteriors(particleDataFrame=results$particleDataFrame,
   priorsMat=results$PriorMatrix,
   realParam=TRUE, realParamValues=c(ancState,genRate))

}
\seealso{
\code{\link{plotPrior}} for a set of functions for visualizing and summarizing
prior and posterior distributions, including a visual comparison for single parameters.
}
\author{
Barb Banbury and Brian O'Meara
}
