% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosteriors.R
\name{plotPosteriors}
\alias{plotPosteriors}
\title{Plot posterior distributions}
\usage{
plotPosteriors(particleDataFrame, priorsMat, realParam = FALSE,
  realParamValues = NA)
}
\arguments{
\item{particleDataFrame}{A \code{particleDataFrame} object returned by TreEvo ABC analyses, can be a
single data frame or a list of data frames.
If the \code{particleDataFrame} is a list of separate TreEvo runs, posteriors will
be layered over each other to check for repeatability.  The relative gray-scale of
posterior distributions in the plot depends on their total number of runs.}

\item{priorsMat}{A \code{PriorMatrix} object s returned by TreEvo \code{\link{doRun}} functions. Can be a single such matrix
or a list of matrices from a series of analyses.
If the \code{PriorMatrix} is a list of matrices, only the first matrix will be
evaluated, and all other matrices will be ignored.
In other words, prior matrices are expected to be identical for all runs considered to the first matrix given.}

\item{realParam}{If \code{TRUE}, this function will plot line segments where real
parameter values are known. (Usually only true when simulated data is analyzed.)}

\item{realParamValues}{Values for real parameters, include a value for each
parameter (including fixed values). Otherwise should be \code{NA}.}
}
\value{
Returns a plot for each free parameter.
}
\description{
For each free parameter in the posterior, this function creates a plot of the distribution of values estimated in the last generation.
This function can also be used to visually compare against true (generating) parameter values in a simulation.
}
\note{
\code{realParam} and \code{realParamValues} should only be used with simulated data, where
the true values are known.
}
\examples{
data(simRunExample)

# make a list of particleDataFrames to plot multiple runs
resultsPDFlist <- lapply(results, function(x) x$particleDataFrame)

plotPosteriors(resultsPDFlist, 
   priorsMat = results[[1]]$PriorMatrix, 
   realParam = TRUE, realParamValues = c(ancState, genRate))
}
\seealso{
\code{\link{plotPrior}} for a set of functions for visualizing and summarizing
prior and posterior distributions, including a visual comparison for single parameters.
}
\author{
Barb Banbury and Brian O'Meara
}
