% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosteriors.R
\name{plotPosteriors}
\alias{plotPosteriors}
\title{Plot posterior distributions}
\usage{
plotPosteriors(particleDataFrame, priorsList, realParam = FALSE,
  realParamValues = NA)
}
\arguments{
\item{particleDataFrame}{A \code{particleDataFrame} object returned by TreEvo ABC analyses, can be a
single data frame or a list of data frames.
If the \code{particleDataFrame} is a list of separate TreEvo runs, posteriors will
be layered over each other to check for repeatability.  The relative gray-scale of
posterior distributions in the plot depends on their total number of runs.}

\item{priorsList}{A \code{priorList} list object, returned by TreEvo \code{\link{doRun}}
functions, where each element is itself a list of length two, composed of the name for the
prior function (e.g. \code{'uniform'}, and a vector of the parameters for that prior distribution). Can be a
single such list, or a list of \code{priorList} lists from a series of analyses (i.e. a list of depth 3).
If the \code{priorList} is a list of lists (if it is found to be a list of depth 3, only the first list will be
evaluated, and all other lists will be ignored.
In other words, priors are expected to be identical for all runs considered by this function, such
the choice of using the list of priors from the first analysis in the list is acceptable for all comparisons.}

\item{realParam}{If \code{TRUE}, this function will plot line segments where real
parameter values are known. (Usually only true when simulated data is analyzed.)}

\item{realParamValues}{Values for real parameters, include a value for each
parameter (including fixed values). Otherwise should be \code{NA}.}
}
\value{
Returns a plot for each free parameter.
}
\description{
For each free parameter in the posterior, this function creates a plot of the distribution of values estimated in the last generation.
This function can also be used to visually compare against true (generating) parameter values in a simulation.
}
\note{
\code{realParam} and \code{realParamValues} should only be used with simulated data, where
the true values are known.
}
\examples{
data(simRunExample)

# make a list of particleDataFrames to plot multiple runs
resultsPDFlist <- lapply(results, function(x) x$particleDataFrame)

plotPosteriors(resultsPDFlist, 
   priorsList = results[[1]]$priorList, 
   realParam = TRUE, realParamValues = c(ancState, genRate))
}
\seealso{
\code{\link{plotPrior}} for a set of functions for visualizing and summarizing
prior and posterior distributions, including a visual comparison for single parameters.
}
\author{
Barb Banbury and Brian O'Meara
}
