% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPriorPost.R
\name{plotPriorPost}
\alias{plotPriorPost}
\alias{plotPrior}
\alias{plotUnivariatePosteriorVsPrior}
\alias{getUnivariatePriorCurve}
\alias{getUnivariatePosteriorCurve}
\title{Plotting, Summarizing and Comparing the Prior and Posterior Distributions}
\usage{
plotPrior(priorFn = match.arg(arg = priorFn, choices = c("fixed",
  "uniform", "normal", "lognormal", "gamma", "exponential"), several.ok =
  FALSE), priorVariables, plotQuants = TRUE, plotLegend = TRUE)

plotUnivariatePosteriorVsPrior(posteriorCurve, priorCurve,
  label = "parameter", trueValue = NULL, ...)

getUnivariatePriorCurve(priorFn, priorVariables, nPoints = 1e+05,
  from = NULL, to = NULL, alpha = 0.8, coda = FALSE,
  verboseMultimodal = TRUE, ...)

getUnivariatePosteriorCurve(acceptedValues, from = NULL, to = NULL,
  alpha = 0.8, coda = FALSE, verboseMultimodal = TRUE, ...)
}
\arguments{
\item{priorFn}{Prior Shape of the distribution; one of either 
"fixed", "uniform", "normal", 
"lognormal", "gamma", or "exponential".}

\item{priorVariables}{Variables needed to describe the shape of the
distribution, dependent on \code{priorFn}:
\describe{
\item{priorFn = "uniform"}{priorVariables = c(min, max)}
\item{priorFn = "normal"}{priorVariables = c(mean, standard deviation)}
\item{priorFn = "lognormal"}{priorVariables = c(mean, standard deviation)}
\item{priorFn = "gamma"}{priorVariables = c(shape, scale)}
\item{priorFn = "exponential"}{priorVariables = c(rate)}
}}

\item{plotQuants}{If \code{TRUE}, plots line segments at the quantiles}

\item{plotLegend}{If \code{TRUE}, plots legend box with quantile values}

\item{posteriorCurve}{Kernal density estimates for the
posterior distribution from \code{getUnivariatePosteriorCurve}.}

\item{priorCurve}{Kernal density estimates for the 
prior distribution from \code{getUnivariatePriorCurve}.}

\item{label}{X label for plot.}

\item{trueValue}{True parameter value, if any such exists and 
is known (usually only true of simulations).}

\item{...}{For \code{getUnivariatePriorCurve} and \code{getUnivariatePosteriorCurve},
this can contain additional arguments passed to \code{\link{density}}, for use in both
calculating the kernal density estimate for finding the curve, and for estimating
the highest density interval. 
A user may want to mess with this to adjust bandwidth, et cetera.
For \code{plotUnivariatePosteriorVsPrior}, this passes additional commands to the initial
call \code{plot}, and thus can set things like a \code{main} plotting title, among
other things.}

\item{nPoints}{Number of points to draw.}

\item{from}{Lower bound, if any. By default this is \code{NULL} and thus ignored.}

\item{to}{Upper bound, if any. By default this is \code{NULL} and thus ignored.}

\item{alpha}{The threshold used for defining the highest density frequency
cut-off. If the highest density interval is applied to a Bayesian MCMC
posterior sample, then the interval is effectively calculated for this
value as a posterior probability density.}

\item{coda}{Default is \code{FALSE}. If \code{TRUE}, unimodal highest density regions will
instead be calculated using \code{\link{HPDinterval}} from package \code{coda}, which is
similar to the function \code{quantile} in that it calculates only a single interval.}

\item{verboseMultimodal}{If \code{TRUE}, the function will print a message
indicating when the inferred highest density interval is discontinuous, and
thus likely reflects that the supplied data is multimodal.}

\item{acceptedValues}{Vector of accepted particle values.}
}
\value{
\code{plotPrior} and \code{plotUnivariatePosteriorVsPrior} 
produce plots of the respective distributions (see above).

\code{getUnivariatePriorCurve} returns a list of x and y density
coordinates, mean, and the highest density intervals (HDI) for
their respective distribution.

\code{getUnivariatePosteriorCurve} does the same for a 
posterior sample of parameter estimates, returning a 
list of x and y density coordinates, mean, and the 
highest posterior density intervals (HPD).
}
\description{
Assorted functions for visualizing and summarizing the
prior and posterior probability distributions associated with ABC analyses.
}
\details{
Function \code{plotPrior} visualizes the shape of various 
prior probability distributions available in TreEvo ABC analyses, 
while \code{getUnivariatePriorCurve} returns density coordinates and 
summary statistics from user-selected prior probability distribution.

Similarly, function \code{getUnivariatePosteriorCurve} returns 
density coordinates and summary statistics from the posterior distribution 
of an ABC analysis.

Both \code{getUnivariatePriorCurve} and 
\code{getUnivariatePosteriorCurve} also calculate the highest 
density intervals for their respective parameters, 
using the function \code{\link{highestDensityInterval}}.

Function \code{plotUnivariatePosteriorVsPrior} plots the 
univariate density distributions from the prior and 
posterior against each other for comparison, along with the 
highest density intervals (HDI) for both.


The summaries calculated from \code{getUnivariatePriorCurve} 
and \code{getUnivariatePosteriorCurve} are used as the input 
for \code{plotUnivariatePosteriorVsPrior}, hence the relationship of 
these functions to each other, and why they are listed together here.
}
\examples{

data(simRunExample)

# examples with plotPrior

plotPrior(
    priorFn = "exponential", 
    priorVariables = c(10))
    
plotPrior(
    priorFn = "normal", 
    priorVariables = c(1, 2))
    
plotPrior(
    priorFn = "gamma", 
    priorVariables = c(2, .2),
    plotQuants = FALSE, 
    plotLegend = FALSE)

# examples of getting density coordinates and
  # summary statistics from distributions

priorKernal <- getUnivariatePriorCurve(
    priorFn = "normal", 
    priorVariables = c(28, 2), 
    nPoints = 100000, 
    from = NULL, 
    to = NULL, 
    alpha = 0.95)

postKernal <- getUnivariatePosteriorCurve(
    acceptedValues = 
        results[[1]]$particleDataFrame$starting_1, 
    from = NULL, 
    to = NULL, 
    alpha = 0.95)

priorKernal
postKernal

# let's compare this (supposed) prior
  # against the posterior in a plot

plotUnivariatePosteriorVsPrior(
    posteriorCurve = postKernal, 
    priorCurve = priorKernal, 
    label = "parameter", 
    trueValue = NULL)

# cool!

}
\seealso{
Highest posterior densities are calculated via
\code{\link{highestDensityInterval}}.

\code{\link{plotPosteriors}} Plots multiple posteriors
against their priors and potential known values.
}
\author{
Brian O'Meara and Barb Banbury
}
