% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSmethods.R
\name{PLSmethods}
\alias{PLSmethods}
\alias{returnPLSModel}
\alias{PLSTransform}
\title{Fitting Univariate Partial Least Squares Models to Free Parameters in ABC}
\usage{
returnPLSModel(trueFreeValuesMatrix, summaryValuesMatrix, validation = "CV",
  scale = TRUE, variance.cutoff = 95)

PLSTransform(summaryValuesMatrix, pls.model)
}
\arguments{
\item{trueFreeValuesMatrix}{Matrix of true free values from simulations.}

\item{summaryValuesMatrix}{Matrix of summary statistics from simulations.}

\item{validation}{Character argument controlling what validation procedure is used by \link{\code{plsr}}.
Default is \code{"CV"} for cross-validation.}

\item{scale}{This argument is passed to \link{\code{plsr}}.  It may be a numeric vector, or logical. If numeric vector, 
the input is scaled by dividing each variable with the corresponding element of scale. 
If \code{scale = TRUE}, the inpus is scaled by dividing each variable by its sample standard deviation. 
If cross-validation is selected (the default for \code{returnPLSModel}),
scaling by the standard deviation is done for every segment.}

\item{variance.cutoff}{Minimum threshold percentage of variance explained for the
number of components included in the final PLS model fit. This value is a 
percentage and must be between 0 and 100. Default is 95%.}

\item{pls.model}{Output from \link{\code{returnPLSModel}}.}
}
\value{
Function \code{returnPLSModel} returns a PLS model, and function \code{PLSTransform} returns transformed summary statistics.
}
\description{
Function \code{returnPLSModel} fits a PLS regression (using \link{\code{plsr}}) individually to each freely varying parameter of a model, unlike
a true multivariate PLS regression. A secondary step than limits the number
of components to those that explain some minimum cumulative percentage
of variance (see argument \code{variance.cutoff}).  For ABC, this seems to result in much
better results, without one parameter dominating the combined variance.
}
\details{
Function \code{PLSTransform} uses results from a Partial Least Squares (PLS) model fit with \code{returnPLSModel} to transform summary values.
}
\examples{

data(simRun)

PLSmodel<-returnPLSModel(trueFreeValuesMatrix,
	  summaryValuesMatrix=summaryValuesMatrix,
   validation="CV", scale=TRUE, variance.cutoff=95)


PLSTransform(summaryValuesMatrix=summaryValuesMatrix, pls.model=PLSmodel)

}
\seealso{
Function \code{returnPLSModel} effectively wraps function \link{\code{plsr}} from package \code{mvr}.
}
\author{
Brian O'Meara and Barb Banbury
}
