% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSRejection.R
\name{PLSRejection}
\alias{PLSRejection}
\title{Partial Least Squares Rejection}
\usage{
PLSRejection(summaryValuesMatrix, trueFreeValuesMatrix, phy, traits,
  abcTolerance, verbose = TRUE, validation = "CV", scale = TRUE,
  variance.cutoff = 95)
}
\arguments{
\item{summaryValuesMatrix}{Matrix of summary statistics from simulations.}

\item{trueFreeValuesMatrix}{Matrix of true free values from simulations.}

\item{phy}{A phylogenetic tree, in package \code{ape}'s \code{phylo} format.}

\item{traits}{data matrix with rownames identical to \code{phy@tip.label}}

\item{abcTolerance}{Proportion of accepted simulations.}

\item{verbose}{option to print progress to screen.}

\item{validation}{Cross Validation procedure for ABC.}

\item{scale}{Scale for \code{pls.model.list}.}

\item{variance.cutoff}{Variance cutoff for \code{pls.model.list}.}
}
\value{
Returns a list of the particle data frame and ABC distances.
}
\description{
This function performs the ABC-rejection analysis using an input simulation
data. Particles are accepted is they fall sufficiently close to the target
data (within the tolerance). Distances are calculated using \code{abcDistance}.
}
\examples{

data(simRun)
PLSRejection(summaryValuesMatrix=results$summaryValuesMatrix,
  trueFreeValuesMatrix=results$trueFreeValuesMatrix,
	 phy=simPhy, traits=simChar, abcTolerance=results$abcTolerance)

}
\author{
Brian O'Meara and Barb Banbury
}
