% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLSTransform.R
\name{PLSTransform}
\alias{PLSTransform}
\title{PLS Transform}
\usage{
PLSTransform(summaryValuesMatrix, pls.model)
}
\arguments{
\item{summaryValuesMatrix}{Matrix of summary statistics from simulations}

\item{pls.model}{results from returnPLSModel}
}
\value{
Returns transformed summary statistics
}
\description{
Uses results from the PLS model to transform summary values
}
\details{
This function uses the scores from the pls model to transform the summary
statistics.
}
\examples{
#PLSTransform(summaryValuesMatrix, pls.model)
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}
\keyword{PLS}
\keyword{PLSTransform}

