% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullFromPrior.R
\name{pullFromPrior}
\alias{pullFromPrior}
\title{Pull value from prior}
\usage{
pullFromPrior(priorValues, priorFn)
}
\arguments{
\item{priorValues}{Matrix with ncol=number of states (characters) at root
and nrow=2 (two parameters to pass to prior distribution)}

\item{priorFn}{Prior shape}
}
\value{
Returns mean
}
\description{
This function retreives random value from variable prior functions
}
\author{
Brian O'Meara and Barb Banbury
}
\references{
O'Meara and Banbury, unpublished
}

