% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPLSModel.R
\name{returnPLSModel}
\alias{returnPLSModel}
\title{Return Partial Least Squares Model}
\usage{
returnPLSModel(trueFreeValuesMatrix, summaryValuesMatrix, validation = "CV",
  scale = TRUE, variance.cutoff = 95)
}
\arguments{
\item{trueFreeValuesMatrix}{Matrix of true free values from simulations}

\item{summaryValuesMatrix}{Matrix of summary statistics from simulations}

\item{validation}{Cross Validation procedure for abc}

\item{scale}{scale for \code{pls.model.list}}

\item{variance.cutoff}{variance cutoff for \code{pls.model.list}}
}
\value{
Returns a PLS model.
}
\description{
This function fits a PLS regression individually to each freely varying parameter of a model, unlike
a true multivariate PLS regression.  For ABC, this seems to result in much
better results, without one parameter dominating the combined variance.
}
\examples{

data(simRun)

 returnPLSModel(trueFreeValuesMatrix,summaryValuesMatrix,
   validation="CV", scale=scale, variance.cutoff=variance.cutoff)

}
\author{
Brian O'Meara and Barb Banbury
}
