% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPLSModel.R
\name{returnPLSModel}
\alias{returnPLSModel}
\title{return PLS Model}
\usage{
returnPLSModel(trueFreeValuesMatrix, summaryValuesMatrix, validation = "CV",
  scale = TRUE, variance.cutoff = 95)
}
\arguments{
\item{trueFreeValuesMatrix}{Matrix of true free values from simulations}

\item{summaryValuesMatrix}{Matrix of summary statistics from simulations}

\item{validation}{Cross Validation procedure for abc}

\item{scale}{scale for pls.model.list}

\item{variance.cutoff}{variance cutoff for pls.model.list}
}
\value{
Returns a pls model
}
\description{
Returns a PLS Model
}
\details{
This function runs a PLS regression on each free param in the model, unlike
a true multivariate PLS regression.  For ABC, this seems to result in much
better results, without one parameter dominating the combined variance.
}
\examples{

#returnPLSModel(trueFreeValuesMatrix,summaryValuesMatrix, validation="CV", scale=scale, variance.cutoff=variance.cutoff)

}
\references{
O'Meara and Banbury, unpublished
}
\author{
Brian O'Meara and Barb Banbury
}
\keyword{PLS}
\keyword{PLSTransform}
\keyword{returnPLSModel}
