% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/treevo-package.R
\docType{data}
\name{simData}
\alias{simChar}
\alias{simData}
\alias{simPhy}
\title{Simulated data}
\format{simPhy in phylo format and simChar in single column matrix}
\description{
Simulated 30-taxon coalescent tree and simulated character data from BM
intrinsic model. Character data was generated using doSimulation.
}
\keyword{datasets}

