\name{simData}
\docType{data}
\alias{simData}
\alias{simPhy}
\alias{simChar}
\title{Simulated data}
\description{
Simulated 30-taxon coalescent tree and simulated character data from BM intrinsic model. Character data was generated using doSimulation.
}

\usage{simData}

\format{simPhy in phylo format and simChar in single column matrix}

\keyword{datasets}