% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatsLong.R
\name{summaryStatsLong}
\alias{summaryStatsLong}
\title{Calculate summary statistics}
\usage{
summaryStatsLong(phy, traits, niter.brown = 25, niter.lambda = 25,
  niter.delta = 25, niter.OU = 25, niter.white = 25,
  do.CI = is.ultrametric(phy))
}
\arguments{
\item{phy}{Tree (Phylogenetic tree in phylo format)}

\item{traits}{data matrix with rownames equal to phy}

\item{niter.brown}{Number of random starts for BM model (min of 2)}

\item{niter.lambda}{Number of random starts for lambda model (min of 2)}

\item{niter.delta}{Number of random starts for delta model (min of 2)}

\item{niter.OU}{Number of random starts for OU model (min of 2)}

\item{niter.white}{Number of random starts for white model (min of 2)}

\item{do.CI}{Use confidence interval? By default, only for ultrametric trees}
}
\value{
Returns a vector of summary statistics
}
\description{
This function creates a vector of summary statistics for TreEvo analysis
}
\details{
Calculates 17 summary statistics from the fitContinuous.hacked function
(brown.lnl, brown.beta, brown.aic, lambda.lnl, lambda.beta, lambda.lambda,
lambda.aic, delta.lnl, delta.beta, delta.delta, delta.aic, ou.lnl, ou.beta,
ou.alpha, ou.aic, white.lnl, white.aic), plus raw.mean, raw.max, raw.min,
raw.var, raw.median, and all tip character values, phylogenetic independent
contrasts, ancestral state reconstruction values, and the range of ancestral
state reconstruction confidence interval.
}
\examples{

#summaryStatsLong(phy, char)

}
\author{
Brian O'Meara and Barb Banbury
}
