\name{sumStatNames}
\alias{sumStatNames}

\title{Naming function for summary stats}


\usage{
sumStatNames(phy)
}

\arguments{
  \item{phy}{Tree (Phylogenetic tree in phylo format)}
}
\details{Creates a vector of summary stat names
}
\value{
	}
\references{O'Meara and Banbury, unpublished}
\author{Brian O'Meara and Barb Banbury}

