% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumStatNames.R
\name{sumStatNames}
\alias{sumStatNames}
\title{Naming function for summary stats}
\usage{
sumStatNames(phy)
}
\arguments{
\item{phy}{Tree (Phylogenetic tree in phylo format)}
}
\description{
This function creates a vector of summary stat names
}
\references{
O'Meara and Banbury, unpublished
}
\author{
Brian O'Meara and Barb Banbury
}
