% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeD.ABCplots.R
\name{ThreeD.ABCplots}
\alias{ThreeD.ABCplots}
\title{3D ABCplots}
\usage{
ThreeD.ABCplots(particleDataFrame, parameter, show.particles = "none",
  plot.parent = FALSE, realParam = FALSE, realParamValues = NA)
}
\arguments{
\item{particleDataFrame}{particleDataFrame output from doRun}

\item{parameter}{column number of parameter of interest from
particleDataFrame}

\item{show.particles}{option to show particles on 3d plot as "none" or as a
function of "weights" or "distance"}

\item{plot.parent}{option to plot lines on the floor of the 3d plot to show
particle parantage}

\item{realParam}{option to display real parameter value as a solid line,
also must give actual value for this (realParamValues).  Note: this should
only be done with simulated data where real param values are recorded}

\item{realParamValues}{Value for (realParam)}
}
\description{
Plot posterior density distribution for each generation in 3d plot window
}
\details{
This opens a new interactive 3d plotting window and plots the posterior
density distribution of accepted particles from each generation.  Several
options are available to add to the plot: plotting particles by weight or
distance, plotting particle parantage, and plotting the real parameter
values (if known).
}
\examples{

#data(simRun)
#ThreeD.ABCplots(particleDataFrame=results$particleDataFrame, parameter=7, show.particles="none", plot.parent=FALSE, realParam=FALSE, realParamValues=NA) 

}
\references{
O'Meara and Banbury, unpublished
}
\author{
Barb Banbury
}
\keyword{ThreeD.ABCplots}
