% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treevo-package.R
\docType{package}
\name{TreEvo-package}
\alias{TreEvo-package}
\alias{treevo}
\title{TreEvo--abc for comparative methods}
\description{
A package for applying Approximate Bayesian Computation to estimating parameters of trait evolution in comparative analyses.
}
\details{
\tabular{ll}{ Package: \tab TreEvo\cr Type: \tab Package\cr Version: \tab
0.3.3\cr Date: \tab 2012-07-02\cr License: \tab GPL\cr }
}
\examples{

breakThisExample
}
\author{
Brian O'Meara, Barb L. Banbury, David W. Bapst

Maintainer: David Bapst <dwbapst@gmail.com>
}
